Compress the gb_rom_values and gba_rom_values tables with ZX0.
