--- include/gb_rom_values/gb_rom_wrapper.h
@@ -1,7 +1,0 @@
-#ifndef GB_ROM_WRAPPER_H
-#define GB_ROM_WRAPPER_H
-
-#include "eng_gb_rom_values.h"
-#include "fre_gb_rom_values.h"
-
-#endif
