+++ tools/data-generator/include/gb_rom_values/gb_rom_values.h
@@ -1,0 +1,11 @@
+#ifndef _GB_ROM_VALUES_H
+#define _GB_ROM_VALUES_H
+
+#include "gb_rom_values/base_gb_rom_struct.h"
+
+extern const struct GB_ROM gb_rom_values_eng[];
+extern const u16 gb_rom_values_eng_size;
+extern const struct GB_ROM gb_rom_values_fre[];
+extern const u16 gb_rom_values_fre_size;
+
+#endif
