+++ tools/data-generator/include/gba_rom_values/gba_rom_values.h
@@ -1,0 +1,24 @@
+#ifndef _GBA_ROM_VALUES_H
+#define _GBA_ROM_VALUES_H
+
+#include "gba_rom_values/base_gba_rom_struct.h"
+
+extern const struct ROM_DATA rom_data_values_eng[];
+extern const u16 rom_data_values_eng_size;
+
+extern const struct ROM_DATA rom_data_values_fre[];
+extern const u16 rom_data_values_fre_size;
+
+extern const struct ROM_DATA rom_data_values_ger[];
+extern const u16 rom_data_values_ger_size;
+
+extern const struct ROM_DATA rom_data_values_ita[];
+extern const u16 rom_data_values_ita_size;
+
+extern const struct ROM_DATA rom_data_values_jpn[];
+extern const u16 rom_data_values_jpn_size;
+
+extern const struct ROM_DATA rom_data_values_spa[];
+extern const u16 rom_data_values_spa_size;
+
+#endif
