+++ tools/data-generator/src/common.cpp
@@ -3,7 +3,7 @@
 #include <cstdio>
 #include <cstring>
 
-void writeTable(const char* output_path, const char *filename, const uint8_t *buffer, size_t buffer_size)
+void writeTable(const char* output_path, const char *filename, const char *buffer, size_t buffer_size)
 {
     char full_path[4096];
     FILE* f;
